EDITOR-#208 merge editor socket to hook
