--- editor/contexts/WebSocketContext.tsx
@@ -1,20 +1,0 @@
-import React, { useState, useEffect, createContext } from "react";
-import EditorSocketAPI from "../websocket/EditorSocketAPI";
-
-export const WebSocketContext = createContext(null);
-
-export default function WebSocket({ children }) {
-  const [webSocket, setWebSocket] = useState(null);
-
-  useEffect(() => {
-    const editorSocket = new EditorSocketAPI();
-    editorSocket.init();
-    setWebSocket(editorSocket);
-  }, []);
-
-  return (
-    <WebSocketContext.Provider value={webSocket}>
-      {children}
-    </WebSocketContext.Provider>
-  );
-}
